+++ src/components/Button/Button.test.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import Button from 'components/Button';
 import Icon from 'components/Icon';
-import { getColor, getHoverColor, getPressedColor } from './utils';
+import { isValidColor, validColors } from './utils';
 import { create } from 'react-test-renderer';
-import viewsPalette from 'theme/palette';
+import { getBtnColor, getHoverColor, getPressedColor } from './styles';
+import palette from 'theme/palette';
 
 describe('Button component', () => {
 	describe('Component', () => {
@@ -53,25 +54,25 @@ describe('Button component', () => {
 		});
 	});
 
-	describe('Utils', () => {
+	describe('Styles', () => {
 		const exampleColors = [
 			{
 				color: 'blue',
-				valueColor: viewsPalette.blue,
-				valueHover: viewsPalette.blueHover,
-				valuePressed: viewsPalette.bluePressed
+				valueColor: palette.blue,
+				valueHover: palette.blueHover,
+				valuePressed: palette.bluePressed
 			},
 			{
 				color: 'black',
-				valueColor: viewsPalette.black,
-				valueHover: viewsPalette.blackHover,
-				valuePressed: viewsPalette.blackPressed
+				valueColor: palette.black,
+				valueHover: palette.blackHover,
+				valuePressed: palette.blackPressed
 			},
 			{
 				color: 'grey',
-				valueColor: viewsPalette.grey,
-				valueHover: viewsPalette.greyHover,
-				valuePressed: viewsPalette.greyPressed
+				valueColor: palette.grey,
+				valueHover: palette.greyHover,
+				valuePressed: palette.greyPressed
 			}
 		];
 
@@ -79,17 +80,31 @@ describe('Button component', () => {
 
 		test('Should returns default colors', () => {
 			exampleInvalidColors.forEach((color) => {
-				expect(getColor(color)).toEqual(viewsPalette.blue);
-				expect(getHoverColor(color)).toEqual(viewsPalette.blueHover);
-				expect(getPressedColor(color)).toEqual(viewsPalette.bluePressed);
+				expect(getBtnColor(color)).toEqual(palette.blue);
+				expect(getHoverColor(color)).toEqual(palette.blueHover);
+				expect(getPressedColor(color)).toEqual(palette.bluePressed);
 			});
 		});
 
 		test('Should returns correct colors', () => {
 			exampleColors.forEach(({ color, valueColor, valueHover, valuePressed }) => {
-				expect(getColor(color)).toEqual(valueColor);
+				expect(getBtnColor(color)).toEqual(valueColor);
 				expect(getHoverColor(color)).toEqual(valueHover);
 				expect(getPressedColor(color)).toEqual(valuePressed);
+			});
+		});
+	});
+
+	describe('Utils', () => {
+		const invalidColors = ['transparentWhite', 'transparentBlack', 'transparentGrey'];
+		test('should return true if color is valid', () => {
+			validColors.forEach((color) => {
+				expect(isValidColor(color)).toBeTruthy();
+			});
+		});
+		test('should return false if color is invalid', () => {
+			invalidColors.forEach((color) => {
+				expect(isValidColor(color)).toBeFalsy();
 			});
 		});
 	});