+++ src/components/Input/styles.js
@@ -9,32 +9,28 @@ const fontSize = typography.size.medium;
 
 const changeColor = (props) => {
 	const { isFocused, isFloating, disabled, error } = props;
-	switch (true) {
-		case disabled:
-			return palette.grey;
-		case error:
-			return palette.statusRed;
-		case isFocused:
-			return palette.blue;
-		case isFloating:
-			return palette.darkGrey;
-		default:
-			return palette.darkGreyPressed;
-	}
+
+	if (disabled) return palette.grey;
+
+	if (error) return palette.statusRed;
+
+	if (isFocused) return palette.blue;
+
+	if (isFloating) return palette.darkGrey;
+
+	return palette.darkGreyPressed;
 };
 
 const hoverAndFocusColor = (props) => {
 	const { isFocused, disabled, error } = props;
-	switch (true) {
-		case disabled:
-			return palette.grey;
-		case error:
-			return palette.statusRed;
-		case isFocused:
-			return palette.blue;
-		default:
-			return palette.black;
-	}
+
+	if (disabled) return palette.grey;
+
+	if (error) return palette.statusRed;
+
+	if (isFocused) return palette.blue;
+
+	return palette.black;
 };
 
 export default {