+++ src/components/index.js
@@ -17,9 +17,10 @@ import Link from './Link';
 import Drawer from './Drawer';
 import ClickAwayListener from './ClickAwayListener';
 import Skeleton from './Skeleton';
+import palette from 'theme/palette';
+import icons from './Icon/icons.json';
 
 export {
-	ErrorBoundary,
 	Avatar,
 	AvatarGroup,
 	Button,
@@ -29,6 +30,7 @@ export {
 	Color,
 	ColorPicker,
 	Drawer,
+	ErrorBoundary,
 	HTML,
 	Icon,
 	Image,
@@ -37,5 +39,7 @@ export {
 	QRCode,
 	Skeleton,
 	Switch,
-	Textarea
+	Textarea,
+	palette,
+	icons
 };